+++ src/inspect_swe/__init__.py
@@ -1,6 +1,6 @@
 from ._claude_code.claude_code import claude_code
 from ._codex_cli.codex_cli import codex_cli
-from ._tools.download import download_agent_binary
+from ._tools.download import cached_agent_binaries, download_agent_binary
 from ._util.sandbox import SandboxPlatform
 
 try:
@@ -13,6 +13,7 @@ __all__ = [
     "claude_code",
     "codex_cli",
     "download_agent_binary",
+    "cached_agent_binaries",
     "SandboxPlatform",
     "__version__",
 ]